[REFACTOR] 서비스명 변경, 로그인과 회원가입 책임 분리
